Code: decouple dns cache from DNS parsing
